pdfdemo/requirements.txt
pillow==11.1.0
PyPDF4==1.27.0
